convert place page to ts
